Create CUDA flag, get rid of CudaEngine
