--- include/reffine/engine/cudaengine.h
@@ -1,38 +1,0 @@
-#ifndef INCLUDE_REFFINE_CUDA_ENGINE_H_
-#define INCLUDE_REFFINE_CUDA_ENGINE_H_
-
-#include "llvm/ExecutionEngine/Orc/JITTargetMachineBuilder.h"
-#include "llvm/IR/DataLayout.h"
-#include "llvm/IR/LLVMContext.h"
-#include "llvm/IR/LegacyPassManager.h"
-#include "llvm/IR/Module.h"
-#include "llvm/MC/TargetRegistry.h"
-#include "llvm/Support/TargetSelect.h"
-
-using namespace std;
-using namespace llvm;
-
-namespace reffine {
-
-class CUDAEngine {
-public:
-    CUDAEngine(Module& llmod) : llmod(llmod) {}
-
-    static CUDAEngine* Init(Module& llmod);
-    void GeneratePTX();
-    void ExecutePTX(void*, int);
-    void PrintPTX();
-    string GetKernelName();
-
-    Module& llmod;
-
-private:
-    string kernel_name;
-    string ptx_str;
-
-    TargetMachine* get_target();
-};
-
-}  // namespace reffine
-
-#endif  // INCLUDE_REFFINE_CUDA_ENGINE_H_
