+++ include/reffine/pass/llvmgen.h
@@ -16,6 +16,12 @@
 #include "llvm/Linker/Linker.h"
 #include "llvm/Support/MemoryBuffer.h"
 #include "llvm/Support/SourceMgr.h"
+#include "llvm/ExecutionEngine/Orc/JITTargetMachineBuilder.h"
+#include "llvm/IR/DataLayout.h"
+#include "llvm/IR/LLVMContext.h"
+#include "llvm/IR/LegacyPassManager.h"
+#include "llvm/MC/TargetRegistry.h"
+#include "llvm/Support/TargetSelect.h"
 #include "reffine/pass/base/irgen.h"
 
 using namespace std;
@@ -42,7 +48,9 @@ public:
         register_vinstrs();
     }
 
+    static string BuildPTX(shared_ptr<Func>, llvm::Module&);
     static void Build(shared_ptr<Func>, llvm::Module&);
+    
 
 private:
     void register_vinstrs();
@@ -86,6 +94,9 @@ private:
 
     llvm::Module& _llmod;
     unique_ptr<llvm::IRBuilder<>> _builder;
+
+    void init_cuda();
+    llvm::TargetMachine *get_target();
 };
 
 }  // namespace reffine
