+++ main.cpp
@@ -11,6 +11,8 @@
 #include <arrow/c/bridge.h>
 
 #include <z3++.h>
+#include <cuda.h>
+#include <cuda_runtime.h>
 
 #include "reffine/ir/node.h"
 #include "reffine/ir/stmt.h"
@@ -25,7 +27,6 @@
 #include "reffine/pass/z3solver.h"
 #include "reffine/pass/llvmgen.h"
 #include "reffine/engine/engine.h"
-#include "reffine/engine/cudaengine.h"
 #include "reffine/arrow/defs.h"
 #include "reffine/builder/reffiner.h"
 
@@ -286,7 +287,7 @@ shared_ptr<Func> basic_transform_kernel()
     loop->exit_cond = _gte(idx, idx_end);
     auto loop_sym = _sym("loop", loop);
 
-    auto foo_fn = _func("foo", loop, vector<Sym>{vec_out_sym, vec_in_sym}, true);
+    auto foo_fn = make_shared<Func>("foo", loop, vector<Sym>{vec_out_sym, vec_in_sym}, SymTable(), true);
     foo_fn->tbl[idx_addr] = idx_alloc;
 
     return foo_fn;
@@ -392,6 +393,76 @@ shared_ptr<Func> tpcds_query9(ArrowTable& table)
     return foo_fn;
 }
 
+#define checkCudaErrors(err) __checkCudaErrors(err, __FILE__, __LINE__)
+static void __checkCudaErrors(CUresult err, const char *filename, int line)
+{
+    assert(filename);
+    if (CUDA_SUCCESS != err) {
+        const char *ename = NULL;
+        const CUresult res = cuGetErrorName(err, &ename);
+        fprintf(stderr,
+                "CUDA API Error %04d: \"%s\" from file <%s>, "
+                "line %i.\n",
+                err, ((CUDA_SUCCESS == res) ? ename : "Unknown"), filename,
+                line);
+        exit(err);
+    }
+}
+
+void execute_ptx(string kernel_name, string ptx_str, void *arg, int len)
+{
+    CUdevice device;
+    CUmodule cudaModule;
+    CUcontext context;
+    CUfunction function;
+
+    checkCudaErrors(cuInit(0));
+    checkCudaErrors(cuDeviceGet(&device, 0));
+    checkCudaErrors(cuCtxCreate(&context, 0, device));
+
+    char name[128];
+    cuDeviceGetName(name, 128, device);
+    std::cout << "Device name: " << name << endl;
+
+    CUdeviceptr d_arr;
+    checkCudaErrors(cuMemAlloc(&d_arr, sizeof(int64_t) * len));
+    checkCudaErrors(cuMemcpyHtoD(d_arr, arg, sizeof(int64_t) * len));
+
+    CUdeviceptr d_arr_out;
+    checkCudaErrors(cuMemAlloc(&d_arr_out, sizeof(int64_t) * len));
+
+    checkCudaErrors(cuModuleLoadData(&cudaModule, ptx_str.c_str()));
+    checkCudaErrors(
+        cuModuleGetFunction(&function, cudaModule, kernel_name.c_str()));
+    cout << "About to run " << kernel_name << " kernel..." << endl;
+
+    int blockDimX = 32;  // num of threads per block
+    int gridDimX = (len + blockDimX - 1) / blockDimX;  // num of blocks
+    void *kernelParams[] = {
+        &d_arr_out,
+        &d_arr,
+    };
+
+    checkCudaErrors(cuLaunchKernel(function, gridDimX, 1, 1, blockDimX, 1, 1,
+                                   0,  // shared memory size
+                                   0,  // stream handle
+                                   kernelParams, NULL));
+
+    checkCudaErrors(cuCtxSynchronize());
+
+    auto arr_out = (int64_t *)malloc(sizeof(int64_t) * len);
+    checkCudaErrors(cuMemcpyDtoH(arr_out, d_arr_out, sizeof(int64_t) * len));
+    cuMemFree(d_arr);
+    cout << "Output from " << kernel_name << " kernel:" << endl;
+    for (int i = 0; i < len; i++) { cout << arr_out[i] << ", "; }
+    cout << endl << endl;
+
+    cuMemFree(d_arr);
+    cuMemFree(d_arr_out);
+    cuModuleUnload(cudaModule);
+    cuCtxDestroy(context);
+}
+
 void test_kernel() {
     /* Test kernel generation and execution*/
     auto fn = basic_transform_kernel();
@@ -403,17 +474,15 @@ void test_kernel() {
     LLVMGen::Build(fn, *llmod);
     jit->Optimize(*llmod);
 
-    std::string output_ptx;
-    auto cuda = CUDAEngine::Init(*llmod);
-    cuda->GeneratePTX();
-    cuda->PrintPTX();
+    auto output_ptx = LLVMGen::BuildPTX(fn, *llmod);
+    cout << "Generated PTX:" << endl << output_ptx << endl;
 
     int len = 1024;
     int64_t* in_array = new int64_t[len];
     for (int i = 0; i < len; i++) {
         in_array[i] = i;
     }
-    cuda->ExecutePTX((int64_t*)(in_array), len);
+    execute_ptx(llmod->getName().str(), output_ptx, (int64_t*)(in_array), len);
 
     return;
 }
