+++ src/CMakeLists.txt
@@ -18,7 +18,6 @@ set(SRC_FILES
     pass/z3solver.cpp
     pass/llvmgen.cpp
     engine/engine.cpp
-    engine/cudaengine.cpp
     builder/reffiner.cpp
 )
 
@@ -46,17 +45,19 @@ add_library(reffine SHARED ${SRC_FILES} ${CMAKE_CURRENT_BINARY_DIR}/vinstr_str.c
 
 target_link_libraries(reffine ${llvm_libs} ${Z3_LIBRARIES})
 target_include_directories(reffine PUBLIC ${LLVM_INCLUDE_DIRS} ${Z3_CXX_INCLUDE_DIRS} ${CMAKE_CURRENT_SOURCE_DIR}/../include)
-target_include_directories(reffine SYSTEM PUBLIC ${CUDA_INCLUDE_DIRS})
 target_compile_options(reffine PRIVATE -Wall -Wextra -pedantic -Werror -Wno-unused-parameter ${Z3_COMPONENT_CXX_FLAGS})
 
-find_package(CUDA REQUIRED)
-include_directories(${CUDA_INCLUDE_DIRS})
-llvm_map_components_to_libnames(llvm_libs
-  NVPTX
-  NVPTXDesc
-  NVPTXInfo
-  NVPTXCodeGen
-)
-target_link_libraries(reffine ${llvm_libs} ${CUDA_CUDA_LIBRARY})
+option(ENABLE_CUDA "Enable CUDA support" ON)
+if(ENABLE_CUDA)
+  target_include_directories(reffine SYSTEM PUBLIC ${CUDA_INCLUDE_DIRS})
+  include_directories(${CUDA_INCLUDE_DIRS})
+  llvm_map_components_to_libnames(llvm_libs
+    NVPTX
+    NVPTXDesc
+    NVPTXInfo
+    NVPTXCodeGen
+  )
+  target_link_libraries(reffine ${llvm_libs} ${CUDA_CUDA_LIBRARY})
 
-include_directories(/usr/local/cuda/include)
+  include_directories(/usr/local/cuda/include)
+endif()
